[MPI DEV] File reorganization
