+++ src/MPI/include/HiChi_MPI.h
@@ -1,0 +1,7 @@
+#pragma once
+
+// Including everything that has to do with mpi abstraction
+#include "MPI_Utilities.h"
+#include "MPI_Topology.h"
+#include "MPI_FieldUtils.h"
+#include "MPI_FieldExchanger.h"
